Add definitions for the meson build tool
https://mesonbuild.com/

Meson is used by a variety of open source projects. I much prefer it
over cmake.

The qxlsx project has a .wrap file and it's own build definition.
The .wrap file is used to expose it's definition to meson.

It's also possible to meson's wrapdb to fetch third party dependencies so we don't
have to bundle them in the future.

I don't expect we'll keep the qxlxs build definition, but rather just use
their cmake build defintions when we migrate to qt6 or perhaps drop the
dependency altogether.

If we do end up keeping it though, meson can call out to the cmake build
defintion, so we can still easily use it.