mark methods that override virtual functions with the override specifier
https://en.cppreference.com/w/cpp/language/override

> Explanation

In a member function declaration or definition, override specifier ensures that the
function is virtual and is overriding a virtual function from a base class. The
program is ill-formed (a compile-time error is generated) if this is not true.

override is an identifier with a special meaning when used after member function
declarators: it's not a reserved keyword otherwise.