Allow GetBytesFrom(Int|Short) to take integral values > 16 bits via concepts
These functions use a new C++ 20 feature called concepts which allow you to specialize templates
more easily.

These are specialized as such to remove a fair amount narrowing warnings in the codebase.
Eventually we do want to solve them correctly, but this means we don't have to cast everywhere in
the meantime.