Make the application version accessible to the program via version.h.in
This works via a generated version.h via a file called version.h.in, in which the version is
substituted by meson itself when you compiled it.

This is pretty standard practice in the open source world.

This currently only works on the meson build system. There's a way to do it with qmake, but i
haven't yet figured it out.