Replace QByteArray with ssd::Buffer
This is a subclass of vector<int8_t> (using private inheritance). It only exposes the methods
currently used in the codebase. If more are needed, then they can be added.

A fair number of of types had to be modified since standard containers
use size_t for size() while QByteArray used int.