Move XLSX processing completely out of Instruction
No Instruction actually modified the xlsx data, so no functionality was lost.

The parsing was moved to Builder, at least temporarily. It just made sense to live alongside read_functions_xlsx
until we decide what to do with that as well.

If we do want to have per Instruction customization, then it should be done in an input/output agnostic way
in which the readers or writers could parse (xlxs, csv, whatever) can parse.